+++ lib/handlers/marshalling/index.ts
@@ -2,6 +2,6 @@ export * from './cached-route-marshaller'
 export * from './cached-routes-marshaller'
 export * from './currency-amount-marshaller'
 export * from './pair-marshaller'
-export * from '../../../marshalling/pool-marshaller'
+export * from './pool-marshaller'
 export * from './route-marshaller'
 export * from './token-marshaller'
