+++ lib/handlers/marshalling/route-marshaller.ts
@@ -2,7 +2,7 @@ import { MixedRoute, V2Route, V3Route } from '@uniswap/smart-order-router/build/
 import { Protocol } from '@uniswap/router-sdk'
 import { MarshalledToken, TokenMarshaller } from './token-marshaller'
 import { MarshalledPair, PairMarshaller } from './pair-marshaller'
-import { MarshalledPool, PoolMarshaller } from '../../../marshalling/pool-marshaller'
+import { MarshalledPool, PoolMarshaller } from './pool-marshaller'
 import { Pool } from '@uniswap/v3-sdk'
 
 export interface MarshalledV2Route {
