+++ lib/handlers/router-entities/route-caching/dynamo-route-caching-provider.ts
@@ -12,7 +12,7 @@ import { Currency, CurrencyAmount, Token, TradeType } from '@uniswap/sdk-core'
 import { Protocol } from '@uniswap/router-sdk'
 import { CACHED_ROUTES_CONFIGURATION } from './cached-routes-configuration'
 import { PairTradeTypeChainId } from './model/pair-trade-type-chain-id'
-import { CachedRoutesMarshaller } from './marshalling/cached-routes-marshaller'
+import { CachedRoutesMarshaller } from '../../marshalling/cached-routes-marshaller'
 import { CachedRoutesStrategy } from './model/cached-routes-strategy'
 import { ProtocolsBucketBlockNumber } from './model/protocols-bucket-block-number'
 import { CachedRoutesBucket } from './model'
