+++ lib/handlers/router-entities/route-caching/index.ts
@@ -1,4 +1,4 @@
-export * from './marshalling'
+export * from '../../marshalling'
 export * from './model'
 export * from './cached-routes-configuration'
 export * from './dynamo-route-caching-provider'
